Words can recurse, using `RECURSE`
Use `RECURSE` to call the word being defined, recursively.

If it's the last call before returning, it's optimized into a BRANCH
back to the beginning of the word (tail recursion optimization.)

Otherwise it generates a _RECURSE instruction, which is followed by
a (negative) offset to the start of the word, like a BRANCH,
but uses `call` so it creates a new call-stack frame.

A word containing a non-optimized _RECURSE has an unknown maximum
stack depth, indicated by setting its StackEffect.max to UINT16_MAX.
Currently there's no special handling for this; it means the runtime
will literally allocate a 65535-deep stack, which is probably enough.
In the future we could put a stack check at the start of every
recursive word, and either abort or grow the stack if necessary.